requirements.txt
# ============================================================
# ML RAG Assessment — requirements.txt
# Python 3.11 | Render-safe pinned versions
# ============================================================

# Core LangChain
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-chroma==0.1.4
langchain-groq==0.1.9
langchain-huggingface==0.0.3

# Vector Store
chromadb==0.5.15

# API
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Evaluation
ragas==0.1.21
datasets==2.21.0

# Embeddings (HuggingFace — free, runs locally)
sentence-transformers==3.0.1

# Utils
python-dotenv==1.0.1
pyyaml==6.0.2
tiktoken==0.7.0

# Numerics — pinned to avoid Render numpy._core errors
numpy==1.26.4
pandas==2.2.2
